Token Update
Dockerfile updated with environment variable for token and Docker presence. Main also updated to check for valid token more effectively.

Will need to come back later to check on dangling aiohttp client_session, aiohttp TCPconnector, and aiohttp ResponseHandler objects

+++ main.py
@@ -4,6 +4,7 @@
 #imports necessary to run the program
 import argparse
 import importlib
+import asyncio
 import os
 import nextcord
 from nextcord.ext import commands
@@ -21,14 +22,7 @@ import config
 Token = config.Token
 extensions = config.extension
 seanToken = config.seanToken
-
-#This allows for the token through be inputted through command line arguments with syntax --token TOKEN
-if args.token != None:
-    configgen.generateConfiguration('m!', False, args.token, seanToken)
-    importlib.reload(config)
-    Token = config.Token
-    extensions = config.extension
-    seanToken = config.seanToken
+newtoken = Token
 
 #Defines the Intents necessary for the bot to communicate with the discord API
 #Also allows the bot to have the permissions needed to run all of its functions
@@ -38,19 +32,59 @@ intents.message_content = True
 
 client = commands.Bot(command_prefix=extensions, intents=intents, help_command=None, case_insensitive=True)
 
-#When the configuration file is generate for the first time, it sets the token to 'TOKEN'
-#This allows the user to define the bot token through stdin
-#Once the token is properly inputted, it deletes and regenerates the configuration file to include the new token for later use
-while len(Token) != 70:
-    print("Invalid Bot Token, please input your Bot Token below")
-    bottoken = input()
-    os.system('clear')
-    os.remove("config.py")
-    configgen.generateConfiguration('m!', False, bottoken, seanToken)
-    importlib.reload(config)
-    Token = config.Token
-    extensions = config.extension
-    seanToken = config.seanToken
+#This grabs any environment variables from something such as docker
+dockerstat = os.environ.get('dockerstatus', False)
+
+#This function allows the user to quickly check if a discord token is valid or not
+async def tokenCheck(token):
+    try:
+        await client.login(token)
+    except nextcord.LoginFailure:
+        await client.close()
+        client.clear()
+        return False
+    return True
+
+#This conditional first checks if it is in a docker container
+#If it is not, then it runs checks for the token and prompts the user to input a different token
+#until it is a correcct token
+if dockerstat:
+    print("Docker Container Detected, Using environment variables instead")
+    newtoken = os.environ.get('token', "TOKEN")
+    checkedtoken = asyncio.run(tokenCheck(newtoken))
+    if not checkedtoken:
+        print("Invalid Token, please try running the container with a different Token")
+        exit(1)
+    Token = newtoken
+else:
+    #This allows for the token through be inputted through command line arguments with syntax --token TOKEN
+    if args.token != None:
+        configgen.generateConfiguration('m!', False, args.token, seanToken)
+        importlib.reload(config)
+        Token = config.Token
+        extensions = config.extension
+        seanToken = config.seanToken
+        newtoken = Token
+
+    checkedtoken = asyncio.run(tokenCheck(newtoken))
+
+    #When the configuration file is generate for the first time, it sets the token to 'TOKEN'
+    #This allows the user to define the bot token through stdin
+    #Once the token is properly inputted, it deletes and regenerates the configuration file to include the new token for later use
+    while not checkedtoken:
+        print("Invalid Bot Token, please input your Bot Token below")
+        newtoken = input()
+        checkedtoken = asyncio.run(tokenCheck(newtoken))
+        os.system('clear')
+        if checkedtoken:
+            os.remove("config.py")
+            configgen.generateConfiguration('m!', False, newtoken, seanToken)
+            importlib.reload(config)
+            Token = config.Token
+            extensions = config.extension
+            seanToken = config.seanToken
+
+os.system('clear')
 
 #This sets the working directory for this section of the program
 pwd = os.path.dirname(os.path.realpath(__file__))
